refactor-app: Shared 패키지 구조 변경
- 도메인별로 아닌, 유형 별로 객체를 나눔
- Enum 등 순수 클래스를 model 파일로 편입
