+++ application/lib/feature/travel/model/create_travel_state.dart
@@ -1,7 +1,6 @@
 import 'package:application_new/feature/geography/model/city_model.dart';
 import 'package:application_new/feature/geography/model/region_model.dart';
-import 'package:application_new/shared/travel/model/travel_companion_type.dart';
-import 'package:application_new/shared/travel/model/travel_motivation.dart';
+import 'package:application_new/shared/model/travel_model.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'create_travel_state.freezed.dart';
