+++ application/lib/feature/travel/provider/create_travel_provider.dart
@@ -4,8 +4,7 @@ import 'package:application_new/feature/authentication/service/auth_service_prov
 import 'package:application_new/feature/geography/model/city_model.dart';
 import 'package:application_new/feature/geography/model/region_model.dart';
 import 'package:application_new/feature/travel/model/create_travel_state.dart';
-import 'package:application_new/shared/travel/model/travel_companion_type.dart';
-import 'package:application_new/shared/travel/model/travel_motivation.dart';
+import 'package:application_new/shared/model/travel_model.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'create_travel_provider.g.dart';
