+++ application/lib/feature/travel_detail/model/travel_detail_model.dart
@@ -1,7 +1,7 @@
 import 'package:application_new/feature/geography/model/city_model.dart';
 import 'package:application_new/feature/travel_detail/model/travel_visit_model.dart';
-import 'package:application_new/shared/place/model/place_model.dart';
-import 'package:application_new/shared/travel/model/travel_model.dart';
+import 'package:application_new/shared/model/place_model.dart';
+import 'package:application_new/shared/model/travel_model.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'travel_detail_model.freezed.dart';
