+++ application/lib/feature/travel_detail/provider/travel_detail_provider.dart
@@ -2,7 +2,7 @@ import 'package:application_new/common/http/http_service_provider.dart';
 import 'package:application_new/feature/travel_detail/model/travel_detail_model.dart';
 import 'package:application_new/feature/travel_detail/model/travel_visit_model.dart';
 import 'package:application_new/feature/travel_detail/provider/travel_detail_state.dart';
-import 'package:application_new/shared/travel/model/travel_model.dart';
+import 'package:application_new/shared/model/travel_model.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'travel_detail_provider.g.dart';
