--- application/lib/shared/member/model/gender.dart
@@ -1,3 +1,0 @@
-enum Gender {
-  male, female, none
-}
