+++ application/lib/shared/model/place_model.dart
