+++ application/lib/shared/model/place_model.freezed.dart
