--- application/lib/shared/travel/model/travel_companion_type.dart
@@ -1,13 +1,0 @@
-enum TravelCompanionType {
-  spouse,
-  children,
-  parents,
-  grandparents,
-  siblings,
-  relatives,
-  friends,
-  lover,
-  colleagues,
-  members,
-  others
-}
