--- application/lib/shared/travel/model/travel_motivation.dart
@@ -1,10 +1,0 @@
-enum TravelMotivation {
-  adventure,
-  rest,
-  friendship,
-  selfReflection,
-  socialNetwork,
-  fitness,
-  newExperiences,
-  education,
-}
